# application/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Programmer Notebook
  version: 1.0.0
servers:
  - url: https://0.0.0.0:8080/api/v2
...
paths:
...
  /registration:
    post:
      tags:
        - auth
      summary: Create user
      operationId: registerUser

      requestBody:
        description: Информация о новом пользователе для регистрации
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/UserRegistrationInfo'
        required: true

      responses:
        "200":
          description: User successfully created
        "409":
          description: User alredy exists
        "422":
          description: Invalid parameters
        "500":
          description: Internal error
  /login:
    post:
      tags:
        - auth
      summary: Login
      operationId: signInUser

      requestBody:
        description: Информация о клиенте для входа
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/UserLoginInfo'
        required: true

      responses:
        "200":
          description: User successfully logged in
          content:
            'json':
              schema:
                $ref: '#/components/schemas/Token'
        "401":
          description: Authorization error
        "404":
          description: Not Found
        "500":
          description: Internal error

      #####################################################################
      # User API
      #####################################################################
...
  /users/{id}:
    get:
      tags:
        - users
      summary: Get user
      description: Получение одного пользователя
      operationId: getUser


      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserFullInfo'
        "401":
          description: Authorization error
        "403":
          description: Недостаточно прав
        "404":
          description: No such user
        "500":
          description: Internal error

    delete:
      tags:
        - users
      summary: Delete user
      description: Удаление одного пользователя
      operationId: deleteUser

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Success
        "401":
          description: Authorization error
        "403":
          description: Недостаточно прав
        "404":
          description: Пользователь не найден
        "500":
          description: Внутренняя ошибка сервера

    patch:
      tags:
        - users
      summary: Update user
      description: Изменение данных пользователя
      operationId: updateUser

      security:
        - JWTAuth: []

      parameters:
        - name: id
          in: path
          description: "ID, или логин, или ФИО изменяемого пользователя"
          required: true
          schema:
            type: integer

      requestBody:
        description: Новые данные пользователя
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/UserPrivateInfo'
        required: true

      responses:
        "200":
          description: Success
          content: {}
        "401":
          description: Authorization error
          content: {}
        "403":
          description: Недостаточно прав
          content: {}
        "404":
          description: No such user
          content: {}
        "500":
          description: Internal error
          content: {}

      #####################################################################
      # Team API
      #####################################################################
...
components:
  schemas:
...
    UserRegistrationInfo:
      required:
        - fio
        - login
        - password
      type: object
      properties:
        fio:
          type: string
        login:
          type: string
        password:
          type: string
    UserLoginInfo:
      required:
        - login
        - password
      type: object
      properties:
        login:
          type: string
        password:
          type: string
...
    UserPrivateInfo:
      required:
        - id
        - fio
        - login
        - password
        - role
      type: object
      properties:
        id:
          type: integer
          format: int
          example: 10
        fio:
          type: string
          example: Иванов Иван Иванович
        login:
          type: string
          example: login
        password:
          type: string
          example: password
        role:
          type: integer
          description: "Роль пользователя (Админ, Читатель, Автор)"
          format: int
          example: 1
    UserFullInfo:
      required:
        - id
        - fio
        - login
        - registrationDate
        - password
        - role
      type: object
      properties:
        id:
          type: integer
          format: int
          example: 10
        fio:
          type: string
          example: Иванов Иван Иванович
        registrationDate:
          type: string
          example: 01.01.2001
        login:
          type: string
          example: login
        password:
          type: string
          example: password
        role:
          type: integer
          description: "Роль пользователя (Админ, Читатель, Автор)"
          format: int
          example: 1

    #####################################################################
    # Team schemas
    #####################################################################
...
    Token:
      type: object
      properties:
        access_key:
          type: string
          description: JWT access key
...
  securitySchemes:
    JWTAuth:
      type: apiKey
      name: token
      in: header